Remove stdio-ldbl.h
The functions which require different alias to long double depending
of the ABI now uses a set os macros that defines the expected alias
on the function prototype, instead of redefine them using the
__LDBL_REDIR_DEC macros.

The  __REDIRECT_LDBL, __REDIRECT_LDBL_NTH, __REDIRECT_LDBL_NTHNL
are refactores to be similar to __REDIRECT with the difference it
now expects the aliases for binary128 and for the long double being
double (which __LDBL_REDIR_DEC does).

All printf internal usage are also replaced to use internal alias
instead.