Remove syslog-ldbl.h
The functions which require different alias to long double depending
of the ABI now uses a set os macros that defines the expected alias
on the function prototype, instead of redefine them using the
__LDBL_REDIR_DEC macros.

All syslog internal usage are also replaced to use internal alias
instead.

Checked on x86_64-linux-gnu and powerpc64le-linux-gnu.